Renamed project grafa_metrics_client to python_metrics_client
