+++ python_metrics_client/exceptions.py
