+++ python_metrics_client/scripts/manage.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import click
-from grafana_celery_client.scripts.grafana import grafana
+from python_metrics_client.scripts.grafana import grafana
 import logging
 import sys
