+++ python_metrics_client/tests/test_metrics_client.py
@@ -1,12 +1,12 @@
 import mock
-from grafana_celery_client.tests.base import TestCase
+from python_metrics_client.tests.base import TestCase
 from datetime import datetime
-from grafana_celery_client.metrics_client import send_metric
+from python_metrics_client.metrics_client import send_metric
 
 
 class MetricsClientTest(TestCase):
 
-    @mock.patch('grafana_celery_client.metrics_client.send_metric_graphite')
+    @mock.patch('python_metrics_client.metrics_client.send_metric_graphite')
     def test_send_metric_graphite(self, graphite_metric_mock):
         timestamp = datetime(2017, 10, 19, 18, 12, 51)
         send_metric('localhost', 2003, 'dev', 'test', 10, [{'product': 'consignado'}], timestamp=timestamp,
@@ -14,7 +14,7 @@ class MetricsClientTest(TestCase):
         graphite_metric_mock.assert_called_with('localhost', 2003, 'dev', 'test', 10, [{'product': 'consignado'}],
                                                 timestamp)
 
-    @mock.patch('grafana_celery_client.metrics_client.send_metric_influx')
+    @mock.patch('python_metrics_client.metrics_client.send_metric_influx')
     def test_send_metric_influx(self, influx_metric_mock):
         timestamp = datetime(2017, 10, 19, 18, 12, 51)
         send_metric('localhost', 8086, 'dev', 'test', 10, [{'product': 'consignado'}], timestamp=timestamp,
